# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FullHauseBot",
    "contact": {
      "name": "Kaduba Egor",
      "email": "egor.kaduba@mail.ru"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/receipts/upload/{max_user_id}": {
      "post": {
        "tags": [
          "receipts"
        ],
        "summary": "Загрузка файла квитанции",
        "operationId": "upload_receipt_receipts_upload__max_user_id__post",
        "parameters": [
          {
            "name": "max_user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Max User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_receipt_receipts_upload__max_user_id__post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/receipts/user/{max_user_id}": {
      "get": {
        "tags": [
          "receipts"
        ],
        "summary": "Получить список квитанций пользователя",
        "operationId": "get_receipt_receipts_user__max_user_id__get",
        "parameters": [
          {
            "name": "max_user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Max User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReceiptsListItem"
                  },
                  "title": "Response Get Receipt Receipts User  Max User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/receipts/{receipt_id}": {
      "get": {
        "tags": [
          "receipts"
        ],
        "summary": "Получить квитанцию по id",
        "operationId": "get_receipt_receipts__receipt_id__get",
        "parameters": [
          {
            "name": "receipt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Receipt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/receipts/check/{receipt_id}": {
      "post": {
        "tags": [
          "receipts"
        ],
        "summary": "Проверка квитанции на ошибки",
        "operationId": "check_receipt_endpoint_receipts_check__receipt_id__post",
        "parameters": [
          {
            "name": "receipt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Receipt Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/receipts/{receipt_id}/complaint": {
      "get": {
        "tags": [
          "receipts"
        ],
        "summary": "генерация жалобы на квитанцию",
        "operationId": "get_complaint_receipts__receipt_id__complaint_get",
        "parameters": [
          {
            "name": "receipt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Receipt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_receipt_receipts_upload__max_user_id__post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_receipt_receipts_upload__max_user_id__post"
      },
...
      "CoefficientResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "service_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Name"
          },
          "coefficient": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coefficient"
          },
          "excess_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excess Amount"
          }
        },
        "type": "object",
        "required": [
          "id",
          "service_name",
          "coefficient",
          "excess_amount"
        ],
        "title": "CoefficientResponse"
      },
      "ErrorSchema": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence",
            "default": "confirmed"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "actual": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual"
          },
          "expected": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected"
          },
          "delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta"
          },
          "legal_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Ref"
          },
          "recommended_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Action"
          }
        },
        "type": "object",
        "required": [
          "type",
          "severity",
          "description"
        ],
        "title": "ErrorSchema"
      },
      "ErrorsResponse": {
        "properties": {
          "receipt_id": {
            "type": "integer",
            "title": "Receipt Id"
          },
          "has_errors": {
            "type": "boolean",
            "title": "Has Errors"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorSchema"
            },
            "type": "array",
            "title": "Errors"
          },
          "checked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Checked At"
          }
        },
        "type": "object",
        "required": [
          "receipt_id",
          "has_errors",
          "errors",
          "checked_at"
        ],
        "title": "ErrorsResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MeterInfoResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "service_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Name"
          },
          "meter_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meter Type"
          },
          "meter_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meter Number"
          },
          "previous_reading": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Reading"
          },
          "current_reading": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Reading"
          },
          "consumption": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consumption"
          }
        },
        "type": "object",
        "required": [
          "id",
          "service_name",
          "meter_type",
          "meter_number",
          "previous_reading",
          "current_reading",
          "consumption"
        ],
        "title": "MeterInfoResponse"
      },
...
      "RecalculationResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "service_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Name"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          }
        },
        "type": "object",
        "required": [
          "id",
          "service_name",
          "reason",
          "amount"
        ],
        "title": "RecalculationResponse"
      },
      "ReceiptDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "period_month": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Month"
          },
          "period_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Year"
          },
          "total_without_insurance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Without Insurance"
          },
          "total_with_insurance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total With Insurance"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "service_charges": {
            "items": {
              "$ref": "#/components/schemas/ServiceChargeResponse"
            },
            "type": "array",
            "title": "Service Charges",
            "default": []
          },
          "meter_infos": {
            "items": {
              "$ref": "#/components/schemas/MeterInfoResponse"
            },
            "type": "array",
            "title": "Meter Infos",
            "default": []
          },
          "coefficients": {
            "items": {
              "$ref": "#/components/schemas/CoefficientResponse"
            },
            "type": "array",
            "title": "Coefficients",
            "default": []
          },
          "recalculations": {
            "items": {
              "$ref": "#/components/schemas/RecalculationResponse"
            },
            "type": "array",
            "title": "Recalculations",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "period_month",
          "period_year",
          "total_without_insurance",
          "total_with_insurance",
          "status"
        ],
        "title": "ReceiptDetailResponse"
      },
      "ReceiptUploadResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "period_month": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Month"
          },
          "period_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Year"
          },
          "total_without_insurance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Without Insurance"
          },
          "total_with_insurance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total With Insurance"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "period_month",
          "period_year",
          "total_without_insurance",
          "total_with_insurance",
          "status"
        ],
        "title": "ReceiptUploadResponse"
      },
      "ReceiptsListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "period_month": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Month"
          },
          "period_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Year"
          },
          "total_without_insurance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Without Insurance"
          },
          "total_with_insurance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total With Insurance"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "period_month",
          "period_year",
          "total_without_insurance",
          "total_with_insurance",
          "status"
        ],
        "title": "ReceiptsListItem"
      },
      "ServiceChargeResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          },
          "service_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Name"
          },
          "volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "tariff": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tariff"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "benefit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Benefit"
          },
          "recalculation": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recalculation"
          },
          "debt_start": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debt Start"
          },
          "paid": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paid"
          },
          "total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "id",
          "section",
          "service_name",
          "volume",
          "unit",
          "tariff",
          "amount",
          "benefit",
          "recalculation",
          "debt_start",
          "paid",
          "total"
        ],
        "title": "ServiceChargeResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}